Allow reassignment of `consume`d variable in try (sometimes)
Prior to this commit, a `consume`d variable could not be
reassigned in a `try` block. This commit changes things to allow
a `consume`d variable to be reassigned within a `try` block only
if it is reassigned in the same expression where it is consumed
(i.e. `z = do_stuff(consume z)`) and only if that expression does
not include any partial calls.