Fix lib/llvm to support OSX
lib/llvm/Makefile: OSX does not support LLVM_USE_LINKER so do not specify
a LLVM_LINKER and instead use the default system linker.

 - Add OSTYPE from the code in Makefile-ponyc
 - Check for greadlink if OSTYPE is osx
 - Default LLVM_LINKER to empty value
 - Only set LLVM_USE_LINKER if LLVM_LINKER is not empty

lib/llvm/llvm-base.rules: OSX does not support the -T parameter so just
use the position parameters. The first parameter is the target
and the second parameter is the name of the symlink, so removing
the -T fixes this incompatibility.