Run docker image validation earlier
It was possible for it not to get run if arm builds failed
and people might merge something broken given that arm
builds fail sometimes.

A more general clean up of circleci ordering is needed but this
is good for now. That more general clean up will be coming
when the release process is changed.