Allow overriding of ponyc version
Nightly builds aren't particularly meaningful with a version like
0.29.0-a334234vv. The version we are telling people is in the form,
nightly-YYYYMMDD. This commit changes our setup to allow the overriding
of the version/tag that is used to identify this build and package
version. It also updates the nightly builds to take advantage of that.

We should as part of the new release process we are moving towards,
revisit the naming we use for all package builds as well as the
displayed version information. This is, however, an excellent start and
should allow us to fairly easily identify nightly releases if folks
report errors and then go get that particular version for testing.

I'm open to the idea that the displayed version should include a SHA like:

nightly-YYYYMMDD-SHA