Slim down CircleCI matrix
We are moving away from supporting system installed LLVM versions
on Linux distros that use glibc. With this move, we can drop a
large number of CI jobs. By dropping those jobs we will be able to:

- Reduce the amount of time it takes to do Linux related CI
- Lower the number of CI images that we need to support

This commit is the first step in that direction. It removes
all CI jobs related to testing system LLVM version on Glibc based
Linux distributions. Additionally, it slims down our "vendored LLVM"
support matrix to a single distro (currently Ubuntu).

A future commit will bring the vendored LLVM testing environment up
to using a more recent version of Linux. This is, however, an excellent
first step.