Fix segfault due to Cycle Detector viewref inconsistency
Prior to this fix, the Cycle Detector could end up in a situation
where it would have an inconsistent viewref which would result
in a segfault with the cycle detector trying to send a message
to an actor that was already destroyed.

This commit fixes the bug by making the cycle detector apply
deltamap updates from actor block messages immediately.

resolves #2977