Add ability to static build ponyc (#3261)
* Use `llvm-config`'s `--system-libs` option
* Add option to statically link ponyc
* Add an alpine static link build

Note: This supersedes https://github.com/ponylang/ponyc/pull/3260 as it changes the hardcoded system libs for llvm to be dynamic based on whatever `llvm-config` outputs. This effectively removes the `ncurses` dependency on linux while leaving it for mac. I haven't updated the docs/dockerfiles in this PR though.

Note the second: the alpine static build for circleci is a temporary hack that needs to be fixed.