Make builds with `musl` on `glibc` systems possible (#3263)
Prior to this commit building with a toolchain based on `musl`
in a `glibc` distribution (like ubuntu) would result in an error.
For these environments `execinfo.h` is not available because
`musl` doesn't include it and these distributions don't package
`libexecinfo` resulting in an error.

This commit changes things to allow the builds to succeed by
ensuring that the `socket.c` includes are only used if the
`__GLIBC__` macro is defined. It also updates `ponyassert.c` to
work around the lack of `execinfo.h` by `ifdef`'ing out the
call to `backtrace` resulting in a binary that compiles but is
not able to print a stacktrace on an assertion.