Stop building AppImages with releases
Currently AppImages fail to be built on release. We are in the process
of changing what and how we release. AppImage(s) based on the download
stats from bintray are little used.

This commit removes AppImage being a supported release format.

We can revisit this as a format in the future once we've sorted out
all our release process and what we support.