Allow fields to be consumed (sometimes) (#3304)
Prior to this commit, fields could not be `consume`d. This commit
changes things to allow fields to be consumed only if they are
reassigned in the same expression where they are `consume`d
(i.e. `z = do_stuff(consume z)` or `z.c = do_stuff(consume z.c)`)
and adhere to the following restrictions:

* only if that expression does not include any partial calls
* consumed fields cannot be referenced in any function calls
  after the field is consumed
* only if that expression does not include any function calls
  on traits or interfaces in any part of the call tree
* the parent object of a field cannot be consumed in that
  expression
* the field cannot be a `let` or `embed` field
* a field of the same type cannot be referenced in any function
  calls after the field is consumed