Build "latest" Docker images on CircleCI
As part of this commit, we need to turn them off being built on DockerHub.

Why don't we want to build on DockerHub?

It's very slow to do there. When releases are done, it takes several hours
to do the various images we need built. Additionally, we doubt that DockerHub
has the ability to build Docker images that use the vendored LLVM that we
are moving to as it takes a long time to complete.

This commit ONLY switches to building "latest" and "alpine" tags on CircleCI.
Release related tags are still done on DockerHub. Before that change can happen,
other changes will need to be made.