Create release binaries and push to Cloudsmith on each release
This is an early preview of where we want to end up release wise. There
is a lot of duplication in the release building and nightly building
scripts. I'm leaving that duplication for now until we have a solid system
in place and then we can look to apply "DRY" principles.