Allow programmatic override of the default runtime options (#3342)
Prior to this commit, the only way to override the default pony
runtime options was via the command line.

This commit allows for programmatic override of the default pony
runtime options via a specific bare function created on the Main
actor:

`fun @runtime_override_defaults(rto: RuntimeOptions) => None`

In order to override the values, a programmer would replace the
`None` with logic to directly set the various options on the `rto`
variable. This bare function is limited in what is allowed and
is only allowed to call functions on `primitive`s to ensure no
memory allocation occurs as it is called before the runtime is
fully initialized.

This commit also renames the internal variable in `options_t` from
`nopin` to `pin` to better align with how it is set from the
command line so it is easier to programmatically change.