Fix building with GCC 8+
Starting with GCC 8, they added a new warning `stringops-truncation`. It
is well intentioned but, seems prone to a lot of false positives.

It found several false positives in the Pony codebase and because we treat
all warnings as errors, you can't build ponyc with GCC 8+.

This commit switches from using strncpy to memcpy. They are in the end, the
same operation except a difference in return type. However, with the new
gcc option, memcpy will not complain about not copying the null terminator.

Fixes #3313