Remove ncurses dependency from Ponyc when using vendored LLVM
The only reason that ponyc needs libncurses installed is because clang
uses libtinfo to decide if it should have colored output or not. And,
for whatever reason, anything that links to LLVM picks up that dependency.

That is unless you compile with `-DLLVM_ENABLE_TERMINFO=OFF`.

With this commit, we will no longer build using TERMINFO when building
LLVM ourselves.

This commit will remove the dependency of nightly builds on libtinfo.

Fixes #3251