Stop using debug ponyc builds for CI
Prior to this commit we were running the following for CI:

release build of pony that:

- built examples
- ran stdlib tests with -d
- ran stdlib tests
- ran the core tests
- validated the grammar

and for the same commit we were doing:

debug build of pony that:

- built examples
- ran stdlib tests with -d
- ran stdlib tests
- ran the core tests
- validated grammar

We decided we weren't getting much out of doing both release and debug build tests of ponyc.
This commit drops the debug ponyc builds. Note, we are still testing the standard library
with the pony specific `--debug` mode. It's only the runtime that we aren't doing debug for.