Stop running any CI on Travis
We are changing our release process and as of the next release
will no longer be releasing packages via COPR nor building .deb
packages that we host on Bintray.

With that change, we no longer need TravisCI. There are some i686
and aarch cross compilation jobs that run there but we are shutting
those down as well.

They have never passed CI and no one seems to use on those platforms,
so... goodbye for now.

If someone wants to come along and support them or otherwise bring up
to date, they can and at that time, we can add CI.