Build documentation on release
Previously, we built the standard library documentation on release
as part of our TravisCI setup. We've since deprecated TravisCI usage.

This commit adds

- building documentation via GitHub Action on each release
- Dockerfile for image that is used for doing the building
- GitHub Action to update the stdlib-builder image daily

This commit remove

- `online-docs` Makefile rule

Although this is set to run on each release, it could be run locally
if we needed to update the documentation. I ran used the image and
script that does the building to update our documentation after the
0.33.0 release failed to build documentation due to an error on the
TravisCI side.

Closes #3362