Add more automation to the release process
Creates a "git ops" driven release process like we do with libraries.
Releasing is triggered by pushing a specially crafted tag like:

release-0.3.1

As part of this commit, the RELEASE_PROCESS documentation has been
updated to reflect the change and the old release script that ran
on the releaser's machine has been removed.

Closes #3272