Turn off changelog bot
- Doesn't work for PRs that came from forks.

Why? Because the triggered event runs in the context of the forked
repo, not the ponyc repo so, it doesn't have access to push the
update.

The basic problem is that a merge of a PR is a PR event and not
a push event. Push events are run in the repo that was merged into.

And the actual rebase/merge is in fact a push event but, it is
disconnected from the PR and we don't have the information we need.