Build latest docker images once a day
As part of our move to using a vendored LLVM that we build for our
ponyc releases, we decided that our Docker images should also use
the vendored LLVM.

However, building the vendored LLVM requires a lot of resources to do
in a timely fashion. None of the available Docker builder environments
are particularly effective at building.

We came together and decided that really, once a day was enough to
update the "latest" images and that using ponyup to install the most
recently nightly builds of ponyc, stable, et al seemed like a very good
approach.

This commit adds a new directory for storing our Dockerfiles that are
used for building these end user prebuilt ponyc images. I am leaving
the old files in place as they are used by both the release and latest
builds and this commit only updates the latest images to use ponyup.

A future commit will add vendored LLVM usage for building release images
and will also remove the then unused Dockerfiles.

This commit partially addresses #3142.