Fix building ponyc with clang on Ubuntu
So Ubuntu and some other Linux distros use the "alternatives" system
to switch common system tools like cc and c++. When a distro uses this
approach, then cc and c++ are symlinks to a program that returns a
value for what compilers to actually use.

So if I want to use clang and clang++ then I can use alternatives to set
cc and c++ accordingly.

Unfortunately, GNU make sets default values for the CC and CXX macros that
do not play well with the alternatives approach.  It sets CC to cc which
plays well with alternatives, but CXX is set to g++ which doesn't.

So on ubuntu, you will get a compilation error if you use alternatives
and try to build ponyc with clang as it will end up using clang and g++.

This commit addresses the situation by checking to see if CC and CXX are
returning conflicting compilers and attempts to address.