Remove custom clone command from CirrusCI setup
At some point in the past when I first tested building lib/llvm
on Cirrus, it required doing a recursive pull of the LLVM submodule.

I wasn't sure why but it was an easy fix. Whatever caused problems in
the past is no longer an issue so we can remove the custom clone
commands.