Trigger latest docker image builds via Cloudsmith upload
Prior to this commit, we were building our daily "latest" Docker images
at a fixed time of day. This should usually be fine as that days nightly
build should be done and uploaded by then.

However, there is no guarantee that this will be the case. It's possible
the daily build could fail or it could be delayed.

This commit changes how latest docker image builds are triggered. Instead
of daily at the same time, we trigger in response to a webhook from the
Cloudsmith "nightlies" repo on package synchronisation.