Move release building from CircleCI to CirrusCI
This is almost the last job to move off of CircleCI for ponyc.
We still have library jobs, but as part of the consolidation
of new release and testing process, we are working our way down
to only using CirrusCI for the heavy lifting jobs and GitHub actions
for everything else.

This is another step in that direction.