Change l suffix to ll for __pony_* builtin wrappers
On Windows, `long` is not a 64-bit wide type like on macOS/Linux and
other *nix systems. The type that's guaranteed to be 64-bit across
the systems targeted by Pony is the `long long` type.

So wrapping the __builtin_*ll macros is more intentional about the input
parameter being a 64-bit integer and will work correctly if we use Clang
to compile Pony on Windows.

This commit also highlights some weirdness (fixed in the next commits):

   #  ifdef PLATFORM_IS_ILP32
   inline uint32_t __pony_ffsll(uint32_t x)
   {
     DWORD i = 0;
     unsigned char non_zero = _BitScanForward(&i, x);
     return non_zero ? i + 1 : 0;
   }
   #  else
   ...
   #  endif

There shouldn't be a version of __pony_ffsll that takes a 32-bit
integer (__pony_ffs is already exactly that).