Add FilePath.symlink test
When working on Windows builds for ponyup, I found that it wouldn't
link due to an error with the FFI call in `symlink`. This commit
adds a test to verify that `symlink` works.

I expect this to fail CI on Windows at which point I will add a fix
and verify that CI then passes on Windows.