Fix linker error when creating symlinks on Windows (#3444)
When working on Windows builds for ponyup, I found that it wouldn't
link due to an error with the FFI call in `symlink`. This commit
adds a test to verify that `symlink` works and fixes said error.

The Windows is ifdef'd out because it requires admin-level privileges to pass and our standard CI is run with normal privileges.