Refactor bit manipulation builtins to correctly handle integer signedness and width (#3443)
For instance, the return type of __pony_ffsl(uint64_t) doesn't have to
be uint64_t because the result bit width doesn't correlate with the
input bit width.

In fact, all the GCC/Clang __builtin_* used here return int. The MSVC
versions use DWORD (unsigned 32-bit) as output parameter type. Hence the
choice of uint32_t as the cross-platform return type for the __pony_*
wrappers.

On Windows, `long` is not a 64-bit wide type like on macOS/Linux and
other *nix systems. The type that's guaranteed to be 64-bit across
the systems targeted by Pony is the `long long` type.

So wrapping the __builtin_*ll macros is more intentional about the input
parameter being a 64-bit integer and will work correctly if we use Clang
to compile Pony on Windows.

There shouldn't be a version of __pony_ffsll that takes a 32-bit
integer (__pony_ffs is already exactly that).

Introduce a third function (__pony_ffszu) that takes a size_t parameter
and chooses if __pony_ffs or __pony_ffsll should be called using the
pre-processor.

This should make writing portable code much easier and fix some
-Wconversion warnings. For instance, this allows us to change the
bitmap_t typedef to be an alias of size_t.

All callsites are passing a 32-bit integer to it, so there's no need to
use the 64-bit version.