Correctly build standard lib docs
So, because of how doc-gen works, you have to be using the correct
compiler version to get the correct documentation version EVEN if
you build from the stdlib of a the correct version of ponyc.

IE checking out ponyc 0.33.2 and building its docs with ponyc 0.30.0,
will get you ponyc docs for version 0.33.0. Because we aren't actually
building the docs. We are building docs for the test suite and we are
using the standard library from version 0.33.0 to build them. Make sense?
I hope so.

This switches our documentation build to not happen until after the following
chain of events:

- release is uploaded to cloudsmith
- event is sent to GitHub
- new release docker image is built
- new stdlib-builder image is built based off the new release docker image
- build the docs and upload using the new stdlib-builder image