Correctly pass PONY_ARCH to CMake
This fixes #3481. CMake was using its own value for the build architecture (`CMAKE_SYSTEM_PROCESSOR`) instead of the one the Makefile was using.  This change passes the value for `arch` that the Makefile is using to CMake to set `PONY_ARCH`.

Thus when `ponyc` is looking for a `../lib/$PONY_ARCH` directory to find `libponyrt.a` in, it'll find the same one that the Makefile created.