Improve cycle detector scheduling by treating it as special actor on the inject queue (#3501)
This commit treats the cycle detector as special system actor that is scheduled
using the global inject queue. Consequently, the cycle detector is subject to work
stealing with a higher priority between all scheduler threads. Prior to this change,
the cycle detector was treated like any other application actor. Moreover, scheduler
thread 0 decided when the cycle detector should be activated for it to potentially do
some work. In some cases, this may have caused pathological behavior, where the time
interval --ponycdinterval for 'poking' the cycle detector was theoretically unbounded.

With this commit, the cycle detector always uses the global non-runtime context (inject_context),
whenever messages are sent to it, independent of the sending actor's context. This guarantees
that it will always be pushed to the global inject queue and will then be pop_global'd by the
first runtime thread to become available.

This has the following effect:

a) get the cycle detector to be scheduled closer to --ponycdinterval
b) have the cycle detector to be scheduled with higher priority as it can be 'pop_global'd' by any scheduler thread from the inject queue.
c) not defer garbage collection of large actor graphs for long running programs right to the end (or close to) a quiescent system.
d) locality is not that important for the cycle detector as for application actors, thus scheduling a previously blocked cycle detector
   using a runtime thread on a first-come-first serve basis should have no negative effect on performance.