Add clang to nightly and release Docker containers
We recently switched to using clang as our default compiler. This
means that if $CC isn't set when compiling a ponyc program it will
look for clang to do linking.

Currently, clang isn't installed in the docker containers nor is $CC
set.

This wasn't an issue previously because gcc was installed so $CC not
being set didn't cause problems.