Make _TestBadExec not depend on file locations.
It was depending on being executed in the directory of its source, which was not true. THe test accidentally created an empty executable version of that file, while ensuring it to be executable. This was polluting the test directory. Now everything is put inside a temp dir, which is cleaned afterwards.