Stop using a magic release branch
It was needed for circle-ci and appveyor workflows, but we no longer use
them so, let's switch to a pure "release from master" as our normal workflow.

The automated usage of a release branch often broken due to odd merge
conflicts on the release branch. That was moderately ok, because they could
be manually fixed and pushed and if you knew the release process well, you
could restart it from the correct point.

However, those problems have pretty much always been caused by the release
branch existing and the merges, so... not so good.

Closes #3533