Fix CommandParser incorrectly handling multiple end-of-option delimiters (#3541)
* Stop swallowing "--" after the first one is processed by command parser

Previously, the command parser would swallow all "--" after the first
one. For me, this manifested itself in not allowing corral to be used
with lldb "corral exec -- lldb ponyc -- $(ponyc) -V=0 -o ./build/
./utility" since the second "--" is required to be passed to lldb.

This change is dependency for a PR being submitted to corral to allow
use of corral with lldb (for debugging ponyc itself).

* adding test for multiple end-of-options issue

* adding test for multiple end-of-options issue (+1 squashed commit)
Squashed commits:
[9ab2aab8] Stop swallowing "--" after the first one is processed by command parser

Previously, the command parser would swallow all "--" after the first
one. For me, this manifested itself in not allowing corral to be used
with lldb "corral exec -- lldb ponyc -- $(ponyc) -V=0 -o ./build/
./utility" since the second "--" is required to be passed to lldb.

This change is dependency for a PR being submitted to corral to allow
use of corral with lldb (for debugging ponyc itself).