Ensure non-blocking process wait and correctly report process exit status (#3419)
* Make waiting for processes non-blocking

by calling waitpid with WNOHANG on posix and doing the equivalent on windows
and if the process didnt exit yet, start a timer and poll the process.
But only create timers etc. if the process didnt finish in the first try to save resources in the normal case.

* Change ProcessNotify.dispose to take a ProcessExitStatus.

So users of ProcessNotify are properly notified of how their process exited. If it was killed by a signal, child_exit_status
will be a Signaled instance containing the signal number that terminated the process (only on POSIX).

* use signals package in test

* fix ponyint_win_process_wait declaration

* include sys/wait.h for WNOHANG on posix

* try fix new process tests on windows.

* use ping -n as a slee preplacement on windows.

* cleanup and simplify to use the existing WaitpidError primitive.

* make WNOHANG primitive private

* update process and process-monitor documentation

* Fix rebase syntax typo.

* Add new file errno.c to libponyrt CMakeLists.txt

* Fix matches on new process error.

* Fix windows test issue

* Re-add _test.sh for process tests.

* Make _TestBadExec not depend on file locations.

It was depending on being executed in the directory of its source, which was not true. THe test accidentally created an empty executable version of that file, while ensuring it to be executable. This was polluting the test directory. Now everything is put inside a temp dir, which is cleaned afterwards.

* Add Manual changelog entry

* Put manual CHANGELOG entries at the right place.

Co-authored-by: Matthias Wahl <matthias@m7w3.de>