Upload Ubuntu 18.04 builds under the correct name
Prior to this change, the Ubuntu 18.04 builds would use an incorrect
triple for building and uploading. They would end up overwriting (if they
came second), the normal gnu/glibc build.

While this change will fix the issue, it shows that it is time to start
making the scripts more compact and starting to DRY them up some. We've
previously discussed during sync calls switching from using bash to python
to make same-ness easier to do. I'm not sure we need to switch to python at
this point, however, it is time to remove some duplication. That will come
in another PR.