Make builders generic across Linux and macOS
This still leaves FreeBSD to be done as well as normalizing nightly
and release builders.

For FreeBSD, the only difference would be gmake instead of make. That
seems like a reasonable thing to normalize but I'm going to spend some
time thinking about the best way to do it that isn't totally hacky.

For combining nightly and release, I don't want to do that in bash and
would wait until we switch to python for these. Having the duplication
between two scripts, nightly and release doesn't seem that bad so I'm
in no hurry to merge them via python rewrite.