Switch Linux nightly and release config to matrix
As we are going to be adding more Linux distros that we support based
on community request, using a matrix of the few differences between
each should make it easier for us to support adding new ones.

We could eventually incorporate the macOS and FreeBSD builds as well,
but this seems like a good first step.