Force removing read-only files on Windows with `FilePath.remove()`
On Windows, `_unlink()` cannot remove a file if it is read-only. This change calls `_chmod(..., _S_IWrite)` on files that are about to be removed so as to make sure they are removable.

This fixes an issue with temp directories being left around from tests in Corral, as Git often makes some of its files read-only.