Change trn->trn to box to fix a soundness hole (#3591)
As discovered in #3571, it is currently possible to violate capability guarantees in the current viewpoint adaptation table, but this could be prevented by updating `ponyc` to use George Steed's viewpoint adaptation rules which distinguish between extracting and non-extracting viewpoint adaptation - (https://github.com/ponylang/ponyc/issues/2815). This commit patches the immediate soundness hole without needing to complete the larger chunk of work.

@sylvanc pointed out that we could easily close the latter hole without implementing the extensive changes required by George Steed's model by simply updating our single viewpoint adaptation table.

This fix will break existing code, in that it might remove some valid/sound uses of non-extracting capabilities to fix the soundness hole for extracting capabilities. But it fixes the hole and prevent the exploit.

To be totally clear, fixing this issue in this fashion means we updated the `trn` row of our viewpoint adaptation table as seen in the tutorial at https://tutorial.ponylang.io/reference-capabilities/combining-capabilities.html to the following:

---

| &#x25B7;        | iso field | trn field | ref field | val field | box field | tag field |
|-----------------|-----------|-----------|-----------|-----------|-----------|-----------|
| __trn origin__  | iso       | box       | box       | val       | box       | tag       |

---

In other words, to close the hole, this commit makes it so that the viewpoint adaptation `trn->trn` would now be a `box` rather than a `trn`.

Note that in George Steed's model, the non-extracting form is `trn` and the extracting form is `val`, so you would expect that we can choose `val` for the "weaker" one but we actually can't guarantee `val` until we know we're extracting, which we don't know because we haven't made that distinction. So the best we can do is have it be `box`.

Partially addresses #3571