Fix building libs with Visual Studio 16.7
Visual Studio 16.7 introduced a change in C++20 compatibility that breaks building LLVM on Windows (#3634).  This change adds a patch for LLVM that fixes the issue.  Fixes #3634.

The change also introduces a check in CMake of the hashes of the patch files. Since the build caches in CI depend on the contents of `libs/CMakeLists.txt`, and that file contains the desired hash, changing the patches (and thus the desired hash) will force a rebuild of the libs build caches (in the same way as changing the Git checkout for LLVM will force a rebuild).