Switch GNU Linux dockerfiles to using Ubuntu 20.04 ponyc
Previously it was based on glibc which might not be the same
as the base image that we are using.

The PR job to test that this image builds is turned off my this
commit as it won't build til after the next release. An issue has
been created to turn it back on after the next release.

Closes #3633