Change the subtyping model used in ponyc to match Steed's model.
The original notion of subtyping in Pony combined subtyping and
aliasing simultaneously (K1 <= K2 if K1^ could be coerced to K2).
This is reasonable for many checks, which used aliasing already,
but this wasn't sufficient for some subtyping checks
(most notably return subtyping, where we do not alias a priori)
and as a result, the subtyping used for return types was unsound,
allowing an `iso` expression (Not `iso^`) to be coerced to any cap,
such as creating a `val` reference while the `iso` was still around.

This change decouples the two, using the subtype ordering for
capabillities which was presented in George Steed's Masters thesis:
"A Principled Design of Capabilites in Pony"

Primarily, this means that most cases where aliasing were required
before now check `K1 <= K2^`, whereas before it was often `!K1 <= K2`
Importantly, this now means that subtyping can be checked
without aliasing by checking that `K1 <= K2`.

Some standard library methods had to be given weaker types due to
the improper return type checks. It may be possible to re-strengthen
these checks with some unsafe code, if it is genuinely sound to do so.

In addition, the tail expression code was strengthened to handle some
cases which were not handled prior. These likely did not arise due
to the incorrect return subtyping checks, but are now necessary to
auto-consume.